# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: string;
  email: string;
  subscriptionTier?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  message: string;
  user: User;
}

export interface Property {
  id: string;
  title: string;
  location: string;
  price: string;
  description: string;
  type: string;
  status: string;
  imageUrl?: string | null;
  createdAt: string;
}

export interface CreatePropertyInput {
  title: string;
  location: string;
  price: string;
  description: string;
  type: string;
  imageUrl?: string | null;
}

export interface Lead {
  id: string;
  fullName: string;
  phone: string;
  status: string;
  createdAt: string;
}

export interface CreateLeadInput {
  fullName: string;
  phone: string;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  features: string[];
  popular: boolean;
}

export interface SuccessResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}