# lib/api-zod/src/generated/types/createLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeadInput {
  fullName: string;
  phone: string;
}