# lib/api-zod/src/generated/types/createPropertyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePropertyInput {
  title: string;
  location: string;
  price: string;
  description: string;
  type: string;
  imageUrl?: string | null;
}