# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: string;
  fullName: string;
  phone: string;
  status: string;
  createdAt: Date;
}