# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  name: string;
  price: number;
  features: string[];
  popular: boolean;
}