# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telangana Estates API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: string;
  title: string;
  location: string;
  price: string;
  description: string;
  type: string;
  status: string;
  imageUrl?: string | null;
  createdAt: Date;
}